# app/static/swagger.yaml
swagger: '2.0'
info:
  title: "Mechanics shop API"
  description: "Used to manage everday functionalites of a Mechanic shop"
  version: "0.0.1"
host: "127.0.0.1:5000" #During Development my API is served on my local host (will change when going to prod)
schemes:
  - "http"
consumes:
  - "application/json" # My API recieves json data
produces:
  - "application/json" # and Here produces a json data
securityDefinitions: #Describing how to add a token auth to the endpoints where its needed/included
  bearerAuth:
    type: apiKey
    name: Authorization
    in: header
paths: #Information about the routes 
  /mechanics/login: #writing out the endpoint URL thats lisitng in postman (the link / whats at the end of the localhost)
    post: #Request type (post,get,put or delete) where the infomatition is gettiong pulled from
      tags: #whatever gets tagged they will pop up under here
        - Mechanics #adds route to a category
      summary: "Login endpoint (Generates a token)" # information about the endpoint (loging in gets you your token)
      description: "Validates mechanics credntials (email and password), and produces a specific token for the mechanic who is getting/ being logged in" #more in-depth information about the endpoint
      parameters: #not all will have a parameter such as Get requests. only when reciecing a body parameter, path parameter, query parameter thats when to add a paramter informtion to the request (info about data being sent! dont forget the s in parameters)
        - in: "body" #where will the info being pulled from/coming from
          name: "body"
          description: "Login Credntials" #whats being sent
          required: true #just lisiting that it will be required 
          schema:
            $ref: "#/definitions/LoginCredentials" #should show that this  is whats being referred  data 
      responses:
        200: #lisiting any response you could recieve 
          description: "Login Successful"
          schema: 
            $ref: "#/definitions/LoginResponse"
          examples: #example of what its going to look like
            application/json:
              message: "successfully logged in."
              status: "successful"
              token: "ex1eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...."
  /mechanics: 
    post: #Create Mechanics Endpoint
      tags:
      - Mechanics
      summary: "Endpoint to Register Mechanic"
      description: "Send in Mechanic info, to create a user profile that will be stored in the data base. DOB and address are opption fields. "
      parameters:
        - in: "body"
          name: "body"
          description: " Mechanic Updates"
          required: true
          schema:
            $ref: "#/definitions/MechanicInput"

      responses:
        201:
          description: "Creation Successful"
          schema:
            $ref: "#/definitions/MechanicResponse"
          examples:
            application/json:
              DOB: "1900-01-01"
              address: "123 Example St."
              email: "test@test1.com"
              id: 0
              password: "Test1"
              first_name: "tester"
              last_name: "Work"
              salary: 45000.0
    
    get: #get all Mechanics Endpoint 
      tags:
        - Mechanics
      summary: "Returns all Mechanics"   
      descriptions: "Endpoint to recieve an array of mechanics objects"
      responses: 
        200:
          description: "Successfully Retrieved Mechanics" 
          schema: 
            $ref: "#/definitions/MechanicResponse"
    
    put: #Update a mechanic
      tags:
        - Mechanics
      summary: "Update Mechanic endpoint"
      description: "Endpoint to update a mechanic, requires a token, mechanic will be identified by the token"
      security:
        - bearerAuth: [] #Referencing our Token Auth definition (Make SUre when authorizing in swagger you add BEarer <Token(like the actual token)> )
      parameters:
        - in: "body"
          name: "body"
          description: "New Mechanic Information"
          required: true
          schema:
            $ref: "#/definitions/MechanicInput"

      responses:
        200:
          description: "Creation Successful"
          schema:
            $ref: "#/definitions/MechanicResponse"
          examples:
            application/json:
              DOB: "1900-01-01"
              address: "123 Example St."
              email: "test@test1.com"
              id: 0
              password: "Test1"
              first_name: "tester"
              last_name: "Work"
              salary: 45000.0

    delete: #Delete a Mechanic
      tags:
        - Mechanics
      summary: "Delete a Mechanic"
      description: "Delete a Mechanic, Requires a token, and that token will be used to identify the mechanic to delete."
      security:
        - bearerAuth: []
      responses: 
        200:
          schema:
            type: object
            properties:
              message:
                type: string
...
definitions: # Information about the shape of the input and output json 
  LoginCredentials: #Input data required for kogin route
    type: object #what has to get sent to login 
    properties:
      email: 
        type: string
      password:
        type: string 

    required:
      - email
      - password 
  LoginResponse: #reciting down the response giving from the body of the terminal of postman from login enpoint(after loging in)
    type: object
    properties: 
      message:
        type: string
      status: 
        type: string
      token:
        type: string
  MechanicInput: #Sending MEechanics Informations
    type: object
    properties:
      DOB:
        type: string
        example: "1900-01-01"
      address:
        type: string
        example: "123 Example St."
      email: 
        type: string
        example: "test@test1.com"
      first_name:
        type: string
        example: "tester"
      id:
        type: integer
        example: "0"
      last_name:
        type: string
        example: "Work"
      password:
        type: string
        example: "Test1"
      salary:
        type: number
        format: float
        example: "1900-01-01"
    
    required:
      - DOB
      - address 
      - email
      - first_name
      - last_name
      - password
  MechanicResponse: #Recieve single MEchanic profile
    type: object 
    properties:
      DOB:
        type: string
      address:
        type: string
      email: 
        type: string
      first_name:
        type: string
      id:
        type: integer
      last_name:
        type: string
      password:
        type: string 
      salary:
        type: number
        format: float
...